+++ extensions/visual-studio-code-recent-projects/src/db.ts
@@ -1,11 +1,11 @@
 import { Alert, Icon, Toast, confirmAlert, showToast } from "@raycast/api";
-import { useSQL } from "@raycast/utils";
+import { executeSQL, useSQL } from "@raycast/utils";
 import fs from "fs";
 import { homedir } from "os";
+import path from "path";
 import { build } from "./preferences";
 import { EntryLike, RecentEntries } from "./types";
-import { isSameEntry } from "./utils";
-import { execFilePromise } from "./utils/exec";
+import { isSameEntry, isWin } from "./utils";
 
 export type RemoveMethods = {
   removeEntry: (entry: EntryLike) => Promise<void>;
@@ -83,11 +83,19 @@ export function useRecentEntries() {
 }
 
 function getPath() {
+  if (isWin) {
+    return path.join(homedir(), `/AppData/Roaming/${build}/User/globalStorage/state.vscdb`);
+  }
+
   return `${homedir()}/Library/Application Support/${build}/User/globalStorage/state.vscdb`;
 }
 
 async function saveEntries(entries: EntryLike[]) {
-  const data = JSON.stringify({ entries });
-  const query = `INSERT INTO ItemTable (key, value) VALUES ('history.recentlyOpenedPathsList', '${data}');`;
-  await execFilePromise("sqlite3", [getPath(), query]);
+  try {
+    const data = JSON.stringify({ entries });
+    const query = `INSERT INTO ItemTable (key, value) VALUES ('history.recentlyOpenedPathsList', '${data}');`;
+    await executeSQL(getPath(), query);
+  } catch (error) {
+    console.error(error);
+  }
 }